test.sh: don’t overwrite test.txt if it already exists

+++ test/test.sh
@@ -7,8 +7,10 @@ rm -f test.txt_ test.new
 
 # We need to copy the source files because the output file will be placed in
 # the same directory. However, $srcdir might not be writable, while $PWD is.
-cp $srcdir/test.txt $srcdir/test.compressed .
-trap 'rm -f test.txt test.compressed' INT TERM EXIT
+if [ ! -e "test.txt" ]; then
+    cp $srcdir/test.txt $srcdir/test.compressed .
+    trap 'rm -f test.txt test.compressed' INT TERM EXIT
+fi
 
 $mscompress_abspath/src/mscompress test.txt
 $mscompress_abspath/src/msexpand < test.txt_ > test.new
